+++ azd/src/main/java/org/azd/common/ApiVersion.java
@@ -47,6 +47,7 @@ public abstract class ApiVersion {
     public static final String WORK = "7.1-preview.1";
     public static final String WORK_ITEM_MIGRATE = "7.1-preview.1";
     public static final String WORK_ITEM_TRACKING = "7.1-preview.3";
+    public static final String WIT_QUERY_BATCH = "7.1-preview.1";
     public static final String WIT_WIQL = "7.1-preview.2";
     public static final String WIT_RECYCLE_BIN = "7.1-preview.2";
     public static final String WORK_ITEM_TYPES = "7.1-preview.2";